Fix the eBPF structure memory layout bug

+++ agent/src/common/lib.rs
@@ -3,7 +3,7 @@
 pub mod protocol;
 mod structs;
 
-pub const MAX_PAYLOAD_SIZE: u32 = 1 << 14;
+pub const MAX_PAYLOAD_SIZE: u64 = 1 << 14;
 pub const TASK_CMD_LEN: usize = 16;
 pub const MAX_PRE_PAYLOAD_SIZE: u32 = 1 << 3;
 pub use protocol::*;
